+++ changelog.markdown
@@ -1,6 +1,7 @@
 ## 1.48.0.0
 
 * Bundled with docu.exe compiled against .Net 4.0.
+* Added targetplatform switch for ILMerge task.
 
 ## 1.46.0.0